LCM-515: - audit logger added in all controllers: - added new class for extracting name/role/origin of the requesting user; - authentication / authorization logs changed to audit logs; - added username in the requests for getting the status of remote LCM; - added username in the requests for exporting users; - fixed create/update method in FetchEndpoint/Metadata services.

+++ lcm-server/src/main/java/nl/kpmg/lcm/server/rest/client/version0/RemoteDataDeletionController.java
@@ -18,8 +18,11 @@ import nl.kpmg.lcm.common.data.User;
 import nl.kpmg.lcm.common.exception.LcmValidationException;
 import nl.kpmg.lcm.common.validation.Notification;
 import nl.kpmg.lcm.server.data.service.RemoteDataDeletionService;
+import nl.kpmg.lcm.server.rest.UserIdentifier;
 import nl.kpmg.lcm.server.rest.authorization.PermissionChecker;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import javax.annotation.security.RolesAllowed;
@@ -44,6 +47,10 @@ import io.swagger.annotations.ApiResponses;
 @Path("client/v0/remoteData/")
 @Api(value = "v0 delete data on Remote LCM")
 public class RemoteDataDeletionController {
+  private static final Logger AUDIT_LOGGER = LoggerFactory.getLogger("auditLogger");
+
+  @Autowired
+  private UserIdentifier userIdentifier;
 
   @Autowired
   private RemoteDataDeletionService service;
@@ -59,22 +66,35 @@ public class RemoteDataDeletionController {
       value = "Metadata Id") @PathParam("metadata_id") final String metadataId, @ApiParam(
       value = "Remote LCM Id") @QueryParam("lcmId") String lcmId) {
 
-    validateRemoteLcmField(metadataId, "Metadata Id");
-    validateRemoteLcmField(lcmId, "Remote Lcm ID");
+    AUDIT_LOGGER.info(userIdentifier.getUserDescription(securityContext, false)
+        + " is trying to delete the data behind the metadata with id: " + metadataId
+        + " on the LCM with id: " + lcmId + ".");
+
+    Notification notification = new Notification();
+    validateRemoteLcmField(metadataId, "Metadata Id", notification);
+    validateRemoteLcmField(lcmId, "Remote Lcm ID", notification);
+    if (notification.hasErrors()) {
+      AUDIT_LOGGER.debug(userIdentifier.getUserDescription(securityContext, true)
+          + " was unable to delete the data behind the metadata with id: " + metadataId
+          + " on the LCM with id: " + lcmId + "." + notification.errorMessage());
+      throw new LcmValidationException(notification);
+    }
 
     User principal = (User) securityContext.getUserPrincipal();
     PermissionChecker.getThreadLocal().set(principal);
     service.deleteRemoteData(lcmId, metadataId);
+    AUDIT_LOGGER.info(userIdentifier.getUserDescription(securityContext, true)
+        + " deleted successfully the data behind the metadata with id: " + metadataId
+        + " on the LCM with id: " + lcmId + ".");
 
     return Response.ok().build();
   }
 
-  private void validateRemoteLcmField(final String field, String fieldName) {
+  private void validateRemoteLcmField(final String field, String fieldName,
+      Notification notification) {
     if (field == null || field.isEmpty() || field.length() > MAX_FIELD_LENGTH) {
-      Notification notification = new Notification();
       notification.addError(fieldName + " could not be null, empty or longer than "
           + MAX_FIELD_LENGTH + "!", null);
-      throw new LcmValidationException(notification);
     }
   }
 }
